backend/pyproject.toml
[project]
name = "fast-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.51.0",
    "beir>=2.1.0",
    "datasets>=3.6.0",
    "faiss-cpu>=1.11.0",
    "fastmcp>=2.3.4",
    "nltk>=3.9.1",
    "pandas>=2.2.3",
    "rank-bm25>=0.2.2",
    "scikit-learn>=1.6.1",
    "seaborn>=0.13.2",
]

[tool.setuptools]
packages = ["trial_mcp_ranking", "trial_mcp_matching", "trial_mcp_retrieval"]
